chore: update metadata.yaml update script to be compatible with Ship.js
